Final Touches
Removed the hover option in the services section as there is no other pages at the moment. Also removed the Project Tittle for the pictures as its not necessary. Added the category on each page as it is necessary. Also modified some the texts in the Why Choose Us Section. Made sure to update the Copyright year. Also added the META Tags so that this page would be searchable.